separate PersistentBehavior API and implementation
