+++ akka-persistence-typed/src/main/scala/akka/persistence/typed/scaladsl/PersistentBehaviors.scala
@@ -3,15 +3,11 @@
  */
 package akka.persistence.typed.scaladsl
 
-import akka.actor.typed.Behavior
 import akka.actor.typed.Behavior.DeferredBehavior
-import akka.actor.typed.internal.TimerSchedulerImpl
-import akka.actor.typed.scaladsl.{ ActorContext, TimerScheduler }
-import akka.annotation.{ DoNotInherit, InternalApi }
-import akka.persistence.{ Recovery, SnapshotSelectionCriteria }
+import akka.actor.typed.scaladsl.ActorContext
+import akka.annotation.InternalApi
+import akka.persistence.SnapshotSelectionCriteria
 import akka.persistence.typed.internal._
-import akka.persistence.typed.scaladsl.PersistentBehaviors.CommandHandler
-import akka.util.ConstantFun
 
 import scala.language.implicitConversions
 
@@ -28,7 +24,7 @@ object PersistentBehaviors {
     initialState:   State,
     commandHandler: (ActorContext[Command], State, Command) ⇒ Effect[Event, State],
     eventHandler:   (State, Event) ⇒ State): PersistentBehavior[Command, Event, State] =
-    new PersistentBehavior(
+    new PersistentBehaviorImpl(
       persistenceId = persistenceId,
       initialState = initialState,
       commandHandler = commandHandler,
@@ -87,69 +83,12 @@ object PersistentBehaviors {
   }
 }
 
-@DoNotInherit
-class PersistentBehavior[Command, Event, State] private (
-  val persistenceId:  String,
-  val initialState:   State,
-  val commandHandler: PersistentBehaviors.CommandHandler[Command, Event, State],
-  val eventHandler:   (State, Event) ⇒ State,
-
-  val recoveryCompleted: (ActorContext[Command], State) ⇒ Unit,
-  val tagger:            Event ⇒ Set[String],
-  val journalPluginId:   String,
-  val snapshotPluginId:  String,
-  val snapshotWhen:      (State, Event, Long) ⇒ Boolean,
-  val recovery:          Recovery
-) extends DeferredBehavior[Command](ctx ⇒
-  TimerSchedulerImpl.wrapWithTimers[Command] { timers ⇒
-    val callbacks = EventsourcedCallbacks[Command, Event, State](
-      initialState,
-      commandHandler,
-      eventHandler,
-      snapshotWhen,
-      recoveryCompleted,
-      tagger
-    )
-    val pluginIds = EventsourcedPluginIds(
-      journalPluginId,
-      snapshotPluginId
-    )
-    new EventsourcedRequestingRecoveryPermit(
-      persistenceId,
-      ctx.asInstanceOf[ActorContext[Any]], // sorry
-      timers.asInstanceOf[TimerScheduler[Any]], // sorry
-      recovery,
-      callbacks,
-      pluginIds
-    ).narrow[Command]
-
-  }(ctx)) {
-
-  def this(
-    persistenceId:  String,
-    initialState:   State,
-    commandHandler: PersistentBehaviors.CommandHandler[Command, Event, State],
-    eventHandler:   (State, Event) ⇒ State) {
-    this(
-      persistenceId,
-      initialState,
-      commandHandler,
-      eventHandler,
-      recoveryCompleted = ConstantFun.scalaAnyTwoToUnit,
-      tagger = (_: Event) ⇒ Set.empty[String],
-      journalPluginId = "" /* default plugin */ ,
-      snapshotPluginId = "" /* default plugin */ ,
-      snapshotWhen = ConstantFun.scalaAnyThreeToFalse,
-      recovery = Recovery()
-    )
-  }
-
+trait PersistentBehavior[Command, Event, State] extends DeferredBehavior[Command] {
   /**
    * The `callback` function is called to notify the actor that the recovery process
    * is finished.
    */
-  def onRecoveryCompleted(callback: (ActorContext[Command], State) ⇒ Unit): PersistentBehavior[Command, Event, State] =
-    copy(recoveryCompleted = callback)
+  def onRecoveryCompleted(callback: (ActorContext[Command], State) ⇒ Unit): PersistentBehavior[Command, Event, State]
 
   /**
    * Initiates a snapshot if the given function returns true.
@@ -158,34 +97,23 @@ class PersistentBehavior[Command, Event, State] private (
    *
    * `predicate` receives the State, Event and the sequenceNr used for the Event
    */
-  def snapshotWhen(predicate: (State, Event, Long) ⇒ Boolean): PersistentBehavior[Command, Event, State] =
-    copy(snapshotWhen = predicate)
-
+  def snapshotWhen(predicate: (State, Event, Long) ⇒ Boolean): PersistentBehavior[Command, Event, State]
   /**
    * Snapshot every N events
    *
    * `numberOfEvents` should be greater than 0
    */
-  def snapshotEvery(numberOfEvents: Long): PersistentBehavior[Command, Event, State] = {
-    require(numberOfEvents > 0, s"numberOfEvents should be positive: Was $numberOfEvents")
-    copy(snapshotWhen = (_, _, seqNr) ⇒ seqNr % numberOfEvents == 0)
-  }
+  def snapshotEvery(numberOfEvents: Long): PersistentBehavior[Command, Event, State]
 
   /**
    * Change the journal plugin id that this actor should use.
    */
-  def withPersistencePluginId(id: String): PersistentBehavior[Command, Event, State] = {
-    require(id != null, "persistence plugin id must not be null; use empty string for 'default' journal")
-    copy(journalPluginId = id)
-  }
+  def withPersistencePluginId(id: String): PersistentBehavior[Command, Event, State]
 
   /**
    * Change the snapshot store plugin id that this actor should use.
    */
-  def withSnapshotPluginId(id: String): PersistentBehavior[Command, Event, State] = {
-    require(id != null, "snapshot plugin id must not be null; use empty string for 'default' snapshot store")
-    copy(snapshotPluginId = id)
-  }
+  def withSnapshotPluginId(id: String): PersistentBehavior[Command, Event, State]
 
   /**
    * Changes the snapshot selection criteria used by this behavior.
@@ -195,37 +123,10 @@ class PersistentBehavior[Command, Event, State] private (
    * You may configure the behavior to skip recovering snapshots completely, in which case the recovery will be
    * performed by replaying all events -- which may take a long time.
    */
-  def withSnapshotSelectionCriteria(selection: SnapshotSelectionCriteria): PersistentBehavior[Command, Event, State] = {
-    copy(recovery = Recovery(selection))
-  }
+  def withSnapshotSelectionCriteria(selection: SnapshotSelectionCriteria): PersistentBehavior[Command, Event, State]
 
   /**
    * The `tagger` function should give event tags, which will be used in persistence query
    */
-  def withTagger(tagger: Event ⇒ Set[String]): PersistentBehavior[Command, Event, State] =
-    copy(tagger = tagger)
-
-  private def copy(
-    initialState:      State                                 = initialState,
-    commandHandler:    CommandHandler[Command, Event, State] = commandHandler,
-    eventHandler:      (State, Event) ⇒ State                = eventHandler,
-    recoveryCompleted: (ActorContext[Command], State) ⇒ Unit = recoveryCompleted,
-    tagger:            Event ⇒ Set[String]                   = tagger,
-    snapshotWhen:      (State, Event, Long) ⇒ Boolean        = snapshotWhen,
-    journalPluginId:   String                                = journalPluginId,
-    snapshotPluginId:  String                                = snapshotPluginId,
-    recovery:          Recovery                              = recovery): PersistentBehavior[Command, Event, State] =
-    new PersistentBehavior[Command, Event, State](
-      persistenceId = persistenceId,
-      initialState = initialState,
-      commandHandler = commandHandler,
-      eventHandler = eventHandler,
-      recoveryCompleted = recoveryCompleted,
-      tagger = tagger,
-      journalPluginId = journalPluginId,
-      snapshotPluginId = snapshotPluginId,
-      snapshotWhen = snapshotWhen,
-      recovery = recovery)
-
+  def withTagger(tagger: Event ⇒ Set[String]): PersistentBehavior[Command, Event, State]
 }
-
